Refs 12540. IReaderDataFilter.hpp moved to interfaces folder.
