+++ include/fastdds/rtps/writer/IReaderDataFilter.hpp
@@ -17,6 +17,6 @@
  */
 
 // This file was moved as part of the implementation of content filtered topics
-#include <fastdds/rtps/IReaderDataFilter.hpp>
+#include <fastdds/rtps/interfaces/IReaderDataFilter.hpp>
 
 FASTDDS_TODO_BEFORE(3, 0, "This header should be removed")
