+++ src/cpp/rtps/writer/StatefulWriter.cpp
@@ -19,7 +19,7 @@
 
 #include <fastdds/rtps/writer/StatefulWriter.h>
 
-#include <fastdds/rtps/IReaderDataFilter.hpp>
+#include <fastdds/rtps/interfaces/IReaderDataFilter.hpp>
 #include <fastdds/rtps/writer/WriterListener.h>
 #include <fastdds/rtps/writer/ReaderProxy.h>
 #include <fastdds/rtps/history/WriterHistory.h>
